% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printify.R
\docType{package}
\name{printify-package}
\alias{printify}
\alias{printify-package}
\title{printify}
\description{
A lightweight message system relying purely on base R. Comes with built-in and
pre styled message types and provides an easy way to create custom messages.
Supports individually styled and colored text as well as timing information.
Designed to make console output more informative and visually organized.
}
\section{Minimal R Version}{

4.1.0 or higher
}

\section{Functions}{
#'
Messages: \code{\link[=print_message]{print_message()}}, \code{\link[=print_headline]{print_headline()}}, \code{\link[=print_start_message]{print_start_message()}}, \code{\link[=print_closing]{print_closing()}},
\code{\link[=print_step]{print_step()}}, \code{\link[=set_up_custom_message]{set_up_custom_message()}}, \code{\link[=get_message_stack]{get_message_stack()}}, \code{\link[=set_no_print]{set_no_print()}},
\code{\link[=print_stack_as_messages]{print_stack_as_messages()}}, \code{\link[=convert_square_brackets]{convert_square_brackets()}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/s3rdia/printify}
  \item \url{https://s3rdia.github.io/printify/}
  \item \url{https://deepwiki.com/s3rdia/printify}
  \item \url{https://s3rdia.github.io/qol_blog/}
}

}
\author{
Tim Siebenmorgen
}
\keyword{internal}
